move LoadSaveFunctionality into AppBar it was only being imported here
